Rework Chapter into a non-abstract generic chapter
The fetcher-specific models (like reddit.Submission) will have an
as_chapter() method that will return a Chapter. Archive and EpubWriter
will deal with Chapter as the standard interface.

Theoretically, this could be done by having the abstract Chapter specify
a method for every field and having the fetcher-specific models implement
them all in their own way.

The advantage of using a concrete Chapter class is that I can instantiate
the same chapter in several ways—in particular, with the archiver. If
Chapter was an abstract interface, then the archiver would need to know
which kind of model (like a reddit.Submission) to build. And
furthermore, reddit.Submission would need multiple constructors. This is
neater.