Make EpubWriter add a barebones title page
It's actually a one-page-long chapter that doesn't appear in the TOC.
I'm not sure if this is the standard EPUB way to implement title pages.
It certainly is not the standard ebooklib way to add covers, which
assumes that the cover will be supplied as an image—but ebooklib seems
to implement that as a chapter just like I did.