Make EPUB output more compliant
Checked with IPDF's EpubCheck tool.

This still does not fully validate. The output file declares EPUB
version 3.0.1, in which `application/x-dtbncx+xml` is a nonstandard
media-type. But that is the type required for the NCX in EPUB 2, and I
want to support readers that only understand version 2. It should be
harmless, since NCX is ignored by EPUB 3 readers.