Implement fetchers.RedditWikiPage
This takes the place of the unfinished RedditMentions, since the completed
functionality of that fetcher wasn't necessary. That "list of IDs" fethcer
was the first I wrote for this whole app, being the simplest method, but not
actually that useful and not compliant with the API that emerged where a
fetcher must provide a get_story() method. If I want to re-implement
that functionality, it should be in a new fetcher.