Set EPUB date metadata
According to the EPUB 3.0 spec, `dc:date` MUST be the date the EPUB
file was generated.

`dcterms:modified` is the date that the content of the book was last
modified.

`dcterms:available` is allowed but not defined in EPUB 3, and is loosely
defined in DCTERMS, but I'm choosing to use it as the date the first
chapter of a story was published.

So, in summary:
  `dcterms:available`   first chapter published
  `dcterms:modified`    most recent chapter published
  `dc:date`             EPUB file generated