Store story_title in every chapter
Some fetchers, like FFNet, are able to automatically find the name of
the story. Save that as a new field on the chapters.

It's denormalized, and is more appropriate to store on the story model.
But the fetchers generate chapter models, and don't have direct access
to the story model. Maybe an opportunity for future refactoring.