Stop using pickled PRAW objects as test fixtures
The object was still making real requests to reddit.com. It's not worth
caputuring and stubbing the requests (with `responses` maybe) or
replaying the HTTP responses with Betamax, because we do have some tests
that hit the real reddit.com (with the LIVE env var set). So we can just
rely on doubles for the unit tests.