Use JSON instead of YAML for archive
The archive was originally intended to be human-readable, but the files
are so large and have so many nested comments that it's not really
working out. I'll write a standalone utility to edit the files with a
standard text editor, or something.

Switch to JSON for less code, more speed, and (slightly) better compatibility.