Move config_loader back to config
Add a get() function that will dereference the global variable in the
config module, which means it will always get the most up-to-date
values.

The get() API differs a little bit from the "standard" in that the
second argument is the type, not a default. Confuse stores defaults in
a separate YAML file.