Move title determination logic out of Story
The "story units" that are fetched are chapters, so we need to infer the
story's title from the chapter data. On some sites, like FFNet, it's
easy because there's a dedicated field. On reddit, it's too difficult
because the story title is embedded in the chapter title.

At any rate, the fetchers should be responsible for determining the
title (or referring to the title supplied on the command line).