Stop storing author and story_title in Chapter
Previously, we were storing those attributes in Chapter because Chapter
represented what was fetched from the remote source. The Story
constructor would extract those values from chapters[0].

This isn't ideal, because different sources store the data differently—
For example, reddit doesn't have a field for story_title. Furthermore,
author is same across the entire story for all supported sources, so
it's not necessary to story it per-chapter.

Instead, store author and title on the Story object. The fetchers (which
instantiate Story) are responsible for extracting the appropriate values
from the Chapters or other sources.