Add `HTMLParser.links_with_rel_value`
This will be useful in the upcoming WordPress fetcher, where "next"
links always (I believe?) have a certain value for the rel attribute,
regardless of the <a> content.

Also do some internal refactoring in HTMLParser. We'll store found links
as bs4 objects instead of namedtuples until after we've filtered them.