Move file reading out of config.py
As part of the effort to split the CLI code from the lib code,
efictopub.config is losing all of its file-reading capability. It will
just take a dict (or dict-like) object with the final config values.

Reading files and parsing command-line args will be the responsibility
of the CLI.