reddit_submission: Cast .edited to an int
The PRAW object will populate .edited with the timestamp of the last
edit, or False if no edits have been made.

In Python 3, False == 0. You can even assert isinstance(False, int).

But we are eventually serializing these values as JSON with jsonpickle,
so let's cast it to an integer just to be extra safe that we get the
expected 0 instead of the string "false".