Rename config `fetch_comments` to `comments`
`comments` is now a mode with three possible values: "all", "none", "author".
It applies to rendering as well as fetching, so it has a more generic name.

For now, the comment pruner call in Chapter depends on a required
fetcher_opt for the author name. Later, we should be able to get the
author name from the fetcher for most fetchers.