Use Confuse to apply fetcher overrides
Since we now have the Confuse object inside Efictopub (instead of just
in cli.opts), we can use its "sources" functionality to add a source for
the fetcher overrides.

The fetcher overrides are copied from the Confuse object itself, and
left there, but the won't do any harm under the "overrides" key.