Add real_live_tests.py
This script runs efictopub end-to-end (as a library) against real live
sites to verify that changes to the sites do not break anything.

A YAML file contains the URLs to fetch and expected values. For
copyright reasons, the expected values are not stored in the git repo.

This script also serves as another example client application.